The_updated_pr

+++ exam/account.php
@@ -6,13 +6,8 @@
 
 <title>AVIATION AND COMMUNICATION</title>
 <link  rel="stylesheet" href="css/bootstrap.min.css"/>
- <link  rel="stylesheet" href="css/bootstrap-theme.min.css"/>    
  <link rel="stylesheet" href="css/main.css">
  <link  rel="stylesheet" href="css/font.css">
- <script src="js/jquery.js" type="text/javascript"></script>
-
- 
-  <script src="js/bootstrap.min.js"  type="text/javascript"></script>
 <link href='http://fonts.googleapis.com/css?family=Roboto:400,700,300' rel='stylesheet' type='text/css'>
  <!--alert message-->
 <?php if(@$_GET['w'])
@@ -43,7 +38,7 @@ $name = $_SESSION['name'];
 $email=$_SESSION['email'];
 
 include_once 'dbConnection.php';
-echo '<span class="pull-right top title1" ><span class="log1"><span class="glyphicon glyphicon-user" aria-hidden="true"></span>&nbsp;&nbsp;&nbsp;&nbsp;Hello,</span> <a href="account.php?q=1" class="log log1">'.$name.'</a>&nbsp;|&nbsp;<a href="logout.php?q=account.php" class="log"><span class="glyphicon glyphicon-log-out" aria-hidden="true"></span>&nbsp;Signout</button></a></span>';
+echo '<span class="pull-right top title1" ><span class="log1"><span class="glyphicon glyphicon-user" aria-hidden="true"></span>&nbsp;&nbsp;&nbsp;&nbsp;Hello,</span> <a href="account.php?q=1" class="log log1">'.$name.'</a>&nbsp;|&nbsp;<a href="logout.php?q=connect.php" class="log"><span class="glyphicon glyphicon-log-out" aria-hidden="true"></span>&nbsp;Signout</button></a></span>';
 }?>
 </div>
 </div></div>
@@ -71,7 +66,7 @@ echo '<span class="pull-right top title1" ><span class="log1"><span class="glyph
 		<li <?php if(@$_GET['q']==3) echo'class="active"'; ?>><a href="account.php?q=3"><span class="glyphicon glyphicon-stats" aria-hidden="true"></span>&nbsp;Ranking</a></li>
 		<li class="pull-right"> <a href="logout.php?q=account.php"><span class="glyphicon glyphicon-log-out" aria-hidden="true"></span>&nbsp;&nbsp;&nbsp;&nbsp;Signout</a></li>
 		</ul>
-            <form class="navbar-form navbar-left" role="search">
+            <form class="navbar-form navbar-left" role="search" onSubmit= "return confirm('You Sure Want to quit the quiz????');" >
         <div class="form-group">
           <input type="text" class="form-control" placeholder="Enter tag ">
         </div>
@@ -85,7 +80,8 @@ echo '<span class="pull-right top title1" ><span class="log1"><span class="glyph
 <div class="col-md-12">
 
 <!--home start-->
-<?php if(@$_GET['q']==1) {
+<?php
+if(@$_GET['q']==1) {
 
 $result = mysqli_query($con,"SELECT * FROM quiz ORDER BY date DESC") or die('Error');
 echo  '<div class="panel"><div class="table-responsive"><table class="table table-striped title1">
@@ -95,10 +91,10 @@ while($row = mysqli_fetch_array($result)) {
 	$title = $row['title'];
 	$total = $row['total'];
 	$sahi = $row['sahi'];
-    $time = $row['time'];
+  $time = $row['time'];
 	$eid = $row['eid'];
 $q12=mysqli_query($con,"SELECT score FROM history WHERE eid='$eid' AND email='$email'" )or die('Error98');
-$rowcount=mysqli_num_rows($q12);	
+$rowcount=mysqli_num_rows($q12);
 if($rowcount == 0){
 	echo '<tr><td>'.$c++.'</td><td>'.$title.'</td><td>'.$total.'</td><td>'.$sahi*$total.'</td><td>'.$time.'&nbsp;min</td>
 	<td><b><a href="account.php?q=quiz&step=2&eid='.$eid.'&n=1&t='.$total.'" class="pull-right btn sub1" style="margin:0px;background:#99cc32"><span class="glyphicon glyphicon-new-window" aria-hidden="true"></span>&nbsp;<span class="title1"><b>Start</b></span></a></b></td></tr>';
@@ -111,22 +107,32 @@ echo '<tr style="color:#99cc32"><td>'.$c++.'</td><td>'.$title.'&nbsp;<span title
 }
 $c=0;
 echo '</table></div></div>';
-
-}?>
+}
+?>
 <span id="countdown" class="timer"></span>
+<?php
+$eid=@$_GET['eid'];
+$result = mysqli_query($con,"SELECT * FROM quiz WHERE eid='$eid'") or die('Error');
+$value = mysqli_fetch_array($result);
+$time = $value['time'];
+$tim=$time*60;
+if(!isset($_COOKIE["time"])) {
+  setcookie("time", time() + $tim, time() + ($tim) + 10, "/");
+}
+?>
 <script>
-var seconds = 75;
+var seconds = <?php echo $_COOKIE["time"]-time(); ?>;
     function secondPassed() {
     var minutes = Math.round((seconds - 30)/60);
     var remainingSeconds = seconds % 60;
     if (remainingSeconds < 10) {
-        remainingSeconds = "0" + remainingSeconds; 
+        remainingSeconds = "0" + remainingSeconds;
     }
     document.getElementById('countdown').innerHTML = minutes + ":" +    remainingSeconds;
-    if (seconds == 0) {
+    if (seconds <= 0) {
         clearInterval(countdownTimer);
         document.getElementById('countdown').innerHTML = "Buzz Buzz";
-    } else {    
+    } else {
         seconds--;
     }
     }
@@ -134,15 +140,59 @@ var countdownTimer = setInterval('secondPassed()', 1000);
 </script>-->
 
 <!--home closed-->
-
 <!--quiz start-->
 <?php
 if(@$_GET['q']== 'quiz' && @$_GET['step']== 2) {
-$eid=@$_GET['eid'];
 $sn=@$_GET['n'];
 $total=@$_GET['t'];
 $q=mysqli_query($con,"SELECT * FROM questions WHERE eid='$eid' AND sn='$sn' " );
 echo '<div class="panel" style="margin:5%">';
+if(mysqli_num_rows(mysqli_query($con,"select * from history where eid= '$eid' and email = '$email'")) == 0) {
+  mysqli_query($con,"INSERT INTO history VALUES('$email','$eid' ,'0','0','0','0',NOW())");
+}
+else if(empty($_SESSION["quizAtm"])) {
+  header("location:update.php?q=quizre&step=25&eid=$eid&n=1&t=$total");
+}
+
+//creating the shortcut to questions
+
+if(@$_GET['sub'] == 'no') {
+  if(!in_array(@$_GET['prev'], $_SESSION["quizAtm"])) {
+    $_SESSION["quizAtm"][] = @$_GET['prev'];
+  }
+}
+
+echo '
+<nav class="navbar navbar-default">
+<button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#quizNav" aria-expanded="false">
+  <span class="sr-only">Toggle navigation</span>
+  OPEN QUSTIONS
+</button>
+<div class="collapse navbar-collapse" id="quizNav">
+  <ul class="nav navbar-nav mr-auto">';
+for ($i=1; $i<=$total; $i++) {
+    if(in_array($i, $_SESSION["quizSub"])) {
+      echo '<li class="nav-item" style="border: 1px solid black; color: white; background-color: green">';
+    }
+    else if(in_array($i, $_SESSION["quizAtm"])) {
+      echo '<li class="nav-item" style="border: 1px solid black; color: white; background-color: yellow">';
+    }
+    else {
+      echo '<li class="nav-item" style="border: 1px solid black; color: white; background-color: red">';
+    }
+    echo'
+        <a class="nav-link quiz" href="account.php?q=quiz&step=2&eid='.$eid.'&n='.$i.'&t='.$total.'&sub=no&prev='.$sn.'">
+          '.$i.'
+        </a>
+      </li>';
+}
+echo '
+  </ul>
+</div>
+</nav>';
+
+//shortcut Created...
+
 while($row=mysqli_fetch_array($q) )
 {
 $qns=$row['qns'];
@@ -150,20 +200,47 @@ $qid=$row['qid'];
 echo '<b>Question &nbsp;'.$sn.'&nbsp;::<br />'.$qns.'</b><br /><br />';
 }
 $q=mysqli_query($con,"SELECT * FROM options WHERE qid='$qid' " );
-echo '<form action="update.php?q=quiz&step=2&eid='.$eid.'&n='.$sn.'&t='.$total.'&qid='.$qid.'" method="POST"  class="form-horizontal">
+
+
+echo '<form id="ansForm" method="post" class="form-horizontal" action="update.php?q=quiz&step=2&fin=no&eid='.$eid.'&n='.$sn.'&t='.$total.'&qid='.$qid.'" >
 <br />';
 
 while($row=mysqli_fetch_array($q) )
 {
-$option=$row['option'];
-$optionid=$row['optionid'];
-echo'<input type="radio" name="ans" value="'.$optionid.'">'.$option.'<br /><br />';
+  $option=$row['option'];
+  $optionid=$row['optionid'];
+  if(array_key_exists($sn, $_SESSION['quizOptions'])) {
+    if($optionid == $_SESSION['quizOptions'][$sn]) {
+      echo'<input type="radio" name="ans"  value="'.$optionid.'" checked>'.$option.' <br /><br />';
+    }
+    else {
+      echo'<input type="radio" name="ans"  value="'.$optionid.'" disabled>'.$option.' <br /><br />';
+    }
+  }
+  else {
+    echo'<input type="radio" name="ans"  value="'.$optionid.'">'.$option.' <br /><br />';
+  }
 }
-echo'<br /><button type="submit" class="btn btn-primary"><span class="glyphicon glyphicon-lock" aria-hidden="true"></span>&nbsp;Submit</button></form></div>';
-header("location:dash.php?q=4&step=2&eid=$id&n=$total");
+echo'<br /><button type="submit" onclick="return subCheckN();" class="btn btn-info">
+<span class="glyphicon glyphicon-forward" aria-hidden="true">
+</span>&nbsp;Next</button>';
+echo'<button type="button" class="btn btn-success" onclick="buttonDisable()">
+<span class="glyphicon glyphicon-ok" aria-hidden="true">
+</span>&nbsp;Confirm</button>';
+echo'<button type="submit" onclick= "return subCheckF();" class="btn btn-primary">
+<span class="glyphicon glyphicon-lock" aria-hidden="true">
+</span>&nbsp;FINAL Submit</button>';
+echo '<button type="reset" onclick="buttonEnable()" class="btn btn-secondry">
+Reset Choices
+</button>
+</form></div>
+';
+header("location:dash.php?q=4&step=2&eid=$eid&n=$total");
+
 }
 //result display
-if(@$_GET['q']== 'result' && @$_GET['eid']) 
+//and reset will reset option locking
+if(@$_GET['q']== 'result' && @$_GET['eid'])
 {
 $eid=@$_GET['eid'];
 $q=mysqli_query($con,"SELECT * FROM history WHERE eid='$eid' AND email='$email' " )or die('Error157');
@@ -177,7 +254,7 @@ $w=$row['wrong'];
 $r=$row['sahi'];
 $qa=$row['level'];
 echo '<tr style="color:#66CCFF"><td>Total Questions</td><td>'.$qa.'</td></tr>
-      <tr style="color:#99cc32"><td>right Answer&nbsp;<span class="glyphicon glyphicon-ok-circle" aria-hidden="true"></span></td><td>'.$r.'</td></tr> 
+      <tr style="color:#99cc32"><td>right Answer&nbsp;<span class="glyphicon glyphicon-ok-circle" aria-hidden="true"></span></td><td>'.$r.'</td></tr>
 	  <tr style="color:red"><td>Wrong Answer&nbsp;<span class="glyphicon glyphicon-remove-circle" aria-hidden="true"></span></td><td>'.$w.'</td></tr>
 	  <tr style="color:#66CCFF"><td>Score&nbsp;<span class="glyphicon glyphicon-star" aria-hidden="true"></span></td><td>'.$s.'</td></tr>';
 }
@@ -187,14 +264,68 @@ while($row=mysqli_fetch_array($q) )
 $s=$row['score'];
 echo '<tr style="color:#990000"><td>Overall Score&nbsp;<span class="glyphicon glyphicon-stats" aria-hidden="true"></span></td><td>'.$s.'</td></tr>';
 }
-echo '</table></div>';
+echo '</table>
+<br>
+<a target= "_blank" href= "view.php?id='.$eid.'">
+<button type="button" class="btn btn-success">Check Answers</button>
+</a>
+</div>
+';
 
 }
 ?>
+
+<!--THIS IS THE CODE FOR OPTION LOCKING -->
+
+
+
+<script>
+var check = false;
+function buttonDisable() {
+  check = true;
+  document.getElementsByName('ans').forEach(function(item) {
+    if(!item.checked) {
+      item.disabled = true;
+    }
+  })
+}
+function subCheckN() {
+  if(check) {
+    return true;
+  }
+  else {
+    document.getElementsByName('ans').forEach(function(item) {
+      if(item.checked) {
+        item.checked = false;
+      }
+    });
+    return confirm('Please confirm your selection before submiting else selection might not be considered valid and might not be evaluated...\nDo you wish to continue submiting(blank answer)???');
+  }
+}
+function subCheckF() {
+  document.getElementById("ansForm").setAttribute("action",<?php echo'"update.php?q=quiz&step=2&fin=yes&eid='.$eid.'&n='.$sn.'&t='.$total.'&qid='.$qid.'"';?>);
+  return confirm('You Sure To Final Submit????');
+}
+function buttonEnable() {
+  check = false;
+  document.getElementsByName('ans').forEach(function(item) {
+    item.disabled = false;
+    if(item.checked) {
+      item.checked= false;
+      item.removeAttribute("checked");
+    }
+  })
+}
+</script>
+
+<!--OPTION LOCKING END -->
+
+
+
 <!--quiz end-->
 <?php
 //history start
-if(@$_GET['q']== 2) 
+if(@$_GET['q']== 2)
 {
 $q=mysqli_query($con,"SELECT * FROM history WHERE email='$email' ORDER BY date DESC " )or die('Error197');
 echo  '<div class="panel title">
@@ -220,7 +351,7 @@ echo'</table></div>';
 }
 
 //ranking start
-if(@$_GET['q']== 3) 
+if(@$_GET['q']== 3)
 {
 $q=mysqli_query($con,"SELECT * FROM rank  ORDER BY score DESC " )or die('Error223');
 echo  '<div class="panel title"><div class="table-responsive">
@@ -264,7 +395,7 @@ echo '</table></div></div>';}
         <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">Close</span></button>
         <h4 class="modal-title" style="font-family:'typo' "><span style="color:orange">Developer</span></h4>
       </div>
-	  
+
       <div class="modal-body">
         <p>
 		<div class="row">
@@ -278,7 +409,7 @@ echo '</table></div></div>';}
 		<h4 style="font-family:'typo' ">jaypee institute of information and technalogy ,Noida .</h4></div></div>
 		</p>
       </div>
-    
+
     </div><!-- /.modal-content
   </div><!-- /.modal-dialog -->
 </div><!-- /.modal -->
@@ -297,7 +428,7 @@ echo '</table></div></div>';}
 <div class="col-md-6">
 <form role="form" method="post" action="admin.php?q=connect.php">
 <div class="form-group">
-<input type="text" name="uname" maxlength="20"  placeholder="Admin user id" class="form-control"/> 
+<input type="text" name="uname" maxlength="20"  placeholder="Admin user id" class="form-control"/>
 </div>
 <div class="form-group">
 <input type="password" name="password" maxlength="15" placeholder="Password" class="form-control"/>
@@ -316,6 +447,34 @@ echo '</table></div></div>';}
 </div><!-- /.modal -->
 <!--footer end-->
 
+<!-- DISABLING ALL LINKS I WHEN IN QUESTIOn PAGE  -->
+
+<?php
+if(@$_GET['q']== 'quiz') {
+  echo '
+  <script>
+    var inputList = Array.prototype.slice.call(document.querySelectorAll("a:not(.quiz)"));
+    inputList.forEach(function(item) {
+      item.setAttribute("onclick", "return confirm(\'You Sure Want to quit the quiz????\');");
+    });
+  </script>
+  ';
+}
+else {
+  $_SESSION["quizSub"] = array();
+  $_SESSION["quizAtm"] = array();
+  $_SESSION["quizOptions"] = array();
+  if(isset($_COOKIE["time"])) {
+    unset($_COOKIE["time"]);
+    setcookie("time", "", time() - 3600, "/");
+  }
+}
+?>
+
+<!-- THATS ALL CODE -->
+
+<script src="js/jquery.js" type="text/javascript"></script>
+<script src="js/bootstrap.min.js"  type="text/javascript"></script>
 
 </body>
 </html>
